moved tests centrally into the tests folder
