+++ tests/test_create_synonym.py
@@ -1,4 +1,4 @@
-from kgcl.tests.test_apply.util import run_test
+from util import run_test
 
 
 def test_create_synonym_with_ids():
