+++ tests/test_delete_edge.py
@@ -1,4 +1,4 @@
-from kgcl.tests.test_apply.util import run_test
+from util import run_test
 
 
 def test_delete_edge_with_ids():
