+++ tests/test_delete_entity.py
@@ -1,4 +1,4 @@
-from kgcl.tests.test_apply.util import run_test
+from util import run_test
 
 
 def test_delete_annotated_edge_with_ids():
